+++ crypto-sodium/lib/Crypto/Encrypt/Public.hs
@@ -38,6 +38,8 @@ module Crypto.Encrypt.Public
   , SecretKey
   , toSecretKey
   , keypair
+  , keypairFromSeed
+  , unsafeKeypairFromSeed
 
   -- * Nonce
   , Nonce
@@ -48,9 +50,17 @@ module Crypto.Encrypt.Public
   , decrypt
   ) where
 
-import Data.ByteArray (ByteArray, ByteArrayAccess)
-import NaCl.Box (Nonce, PublicKey, SecretKey, keypair, toNonce, toPublicKey, toSecretKey)
+import Data.ByteArray (ByteArray, ByteArrayAccess, ScrubbedBytes, withByteArray)
+import Data.ByteArray.Sized as Sized (SizedByteArray, alloc, allocRet)
+import Data.ByteString (ByteString)
+import Data.Functor (void)
+import Data.Proxy (Proxy(..))
+import System.IO.Unsafe (unsafePerformIO)
 
+import qualified Libsodium as Na
+
+import NaCl.Box
+  (Nonce, PublicKey, SecretKey, keypair, toNonce, toPublicKey, toSecretKey)
 import qualified NaCl.Box as NaCl.Box
 
 
@@ -111,3 +121,34 @@ decrypt
   -> ctBytes -- ^ Encrypted message (cyphertext)
   -> Maybe ptBytes
 decrypt = NaCl.Box.open
+
+
+-- | Seed for deterministically generating a keypair.
+--
+-- In accordance with Libsodium's documentation, the seed must be of size
+-- @Na.CRYPTO_BOX_SEEDBYTES@.
+--
+-- This type is parametrised by the actual data type that contains
+-- bytes. This can be, for example, a @ByteString@.
+type Seed a = SizedByteArray Na.CRYPTO_BOX_SEEDBYTES a
+
+
+-- | Generate a new 'SecretKey' together with its 'PublicKey' from a given seed.
+keypairFromSeed
+  :: ByteArrayAccess seed
+  => Seed seed
+  -> IO (PublicKey ByteString, SecretKey ScrubbedBytes)
+keypairFromSeed seed = do
+  allocRet Proxy $ \skPtr ->
+    alloc $ \pkPtr ->
+    withByteArray seed $ \sdPtr ->
+    -- always returns 0, so we don’t check it
+    void $ Na.crypto_box_seed_keypair pkPtr skPtr sdPtr
+
+-- | Generate a new 'SecretKey' together with its 'PublicKey' from a given seed,
+-- in a pure context.
+unsafeKeypairFromSeed
+  :: ByteArrayAccess seed
+  => Seed seed
+  -> (PublicKey ByteString, SecretKey ScrubbedBytes)
+unsafeKeypairFromSeed = unsafePerformIO . keypairFromSeed